# Improve performance of serialization to string and `Display`