```
F128p18/serialize 123.456 to string
                        time:   [49.155 ns 49.222 ns 49.296 ns]
                        change: [-83.871% -83.838% -83.798%] (p = 0.00 < 0.05)
                        Performance has improved.
F128p18/serialize MAX to string
                        time:   [73.920 ns 73.976 ns 74.033 ns]
                        change: [-46.828% -46.714% -46.609%] (p = 0.00 < 0.05)
                        Performance has improved.
F64p9/serialize 123.456 to string
                        time:   [25.929 ns 25.966 ns 26.011 ns]
                        change: [-70.228% -70.153% -70.078%] (p = 0.00 < 0.05)
                        Performance has improved.
F64p9/serialize MAX to string
                        time:   [31.490 ns 31.514 ns 31.541 ns]
                        change: [-64.892% -64.804% -64.702%] (p = 0.00 < 0.05)
                        Performance has improved.
```